fix: `PlayerBlockBreakListener`へマナ切れブロック破壊停止設定取得処理を移動
